+++ lib/dart_pre_commit.dart
@@ -3,7 +3,6 @@ export 'src/hooks_provider.dart';
 export 'src/repo_entry.dart';
 export 'src/task_base.dart';
 export 'src/tasks/analyze_task.dart';
-export 'src/tasks/fix_imports_task.dart';
 export 'src/tasks/format_task.dart';
 export 'src/tasks/outdated_task.dart';
 export 'src/tasks/pull_up_dependencies_task.dart';