+++ test/integration/integration_test.dart
@@ -155,84 +155,9 @@ void main() {
     await testDir.delete(recursive: true);
   });
 
-  test('fix imports', () async {
-    await _git(const ['add', 'lib/src/fix_imports.dart']);
-    await _sut(const ['--no-format', '--no-analyze']);
-
-    final data = await _readFile('lib/src/fix_imports.dart');
-    expect(
-      data,
-      '''
-import 'dart:io';
-
-import 'package:stuff/stuff.dart';
-
-// this is important
-import '../test_project.dart';
-
-void main() {}
-''',
-    );
-  });
-
-  testWithData<Tuple2<String, String>>(
-    'library imports',
-    const [
-      Tuple2('package:test_project/test_project.dart', 'absolute'),
-      Tuple2('package:test_project/another.dart', 'absolute'),
-      Tuple2('../../test_project.dart', 'relative'),
-      Tuple2('../../another.dart', 'relative'),
-    ],
-    (fixture) async {
-      await _writeFile('lib/src/extra/extra.dart', '');
-      await _writeFile('lib/test_project.dart', '');
-      await _writeFile('lib/another.dart', '');
-      await _writeFile(
-        'lib/src/stuff/library_imports.dart',
-        '''
-// this is important
-import '../extra/extra.dart';
-import '${fixture.item1}';
-import 'dart:io';
-import 'package:stuff/test_project.dart';
-// dart_pre_commit:ignore-library-import
-import 'package:test_project/xxx.dart';
-
-void main() {}
-''',
-      );
-
-      await _git(const ['add', 'lib/src/stuff/library_imports.dart']);
-
-      final lines = <String>[];
-      final code = await _sut(
-        const [
-          '--no-fix-imports',
-          '--library-imports',
-          '--no-format',
-          '--no-analyze',
-          '--detailed-exit-code',
-        ],
-        failOnError: false,
-        onStdout: (stream) => stream
-            .transform(utf8.decoder)
-            .transform(const LineSplitter())
-            .listen((line) => lines.add(line)),
-      );
-      expect(code, HookResult.rejected.index);
-      expect(
-        lines,
-        contains(
-          '  [INF] Found ${fixture.item2} import of non-src library: '
-          '${fixture.item1}',
-        ),
-      );
-    },
-  );
-
   test('format', () async {
     await _git(const ['add', 'bin/format.dart']);
-    await _sut(const ['--no-fix-imports', '--no-analyze']);
+    await _sut(const ['--no-analyze']);
 
     final data = await _readFile('bin/format.dart');
     expect(
@@ -253,7 +178,7 @@ void main() {
 
     final lines = <String>[];
     final code = await _sut(
-      const ['--no-fix-imports', '--no-format', '--detailed-exit-code'],
+      const ['--no-format', '--detailed-exit-code'],
       failOnError: false,
       onStdout: (stream) => stream
           .transform(utf8.decoder)
@@ -277,7 +202,6 @@ void main() {
     final lines = <String>[];
     final code = await _sut(
       const [
-        '--no-fix-imports',
         '--no-format',
         '--no-analyze',
         '--check-pull-up',
@@ -297,7 +221,6 @@ void main() {
     final lines = <String>[];
     final code = await _sut(
       const [
-        '--no-fix-imports',
         '--no-format',
         '--no-analyze',
         '--outdated=any',
@@ -314,31 +237,6 @@ void main() {
       lines,
       contains(
         startsWith('  [INF] Required:    mobx: 1.1.0 -> '),
-      ),
-    );
-  });
-
-  test('nullsafe', () async {
-    final lines = <String>[];
-    final code = await _sut(
-      const [
-        '--no-fix-imports',
-        '--no-format',
-        '--no-analyze',
-        '--nullsafe',
-        '--detailed-exit-code',
-      ],
-      failOnError: false,
-      onStdout: (stream) => stream
-          .transform(utf8.decoder)
-          .transform(const LineSplitter())
-          .listen((line) => lines.add(line)),
-    );
-    expect(code, HookResult.rejected.index);
-    expect(
-      lines,
-      contains(
-        startsWith('  [INF] Upgradeable: mobx: 1.1.0 -> '),
       ),
     );
   });