+++ CHANGELOG.md
@@ -4,10 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased - MAJOR] - 20XX-XX-XX
+### Added
+### Changed
+### Deprecated
+### Removed
+- fix imports task
+- library imports task
+- nullsafety task
+### Fixed
+### Security
+
+
 ## [2.3.3] - 2021-07-01
 ### Changed
 - Dependency updates
-
 
 ## [2.3.2] - 2021-07-01
 ### Fixed
@@ -109,7 +120,7 @@ higher version from the lockfile
 ## [0.1.0] - 2020-10-02
 - Initial release
 
-## [Unreleased]
+## [Unreleased] - 20XX-XX-XX
 ### Added
 ### Changed
 ### Deprecated